can now create hallways between rooms, rearranged project directory, added struct for level
