+++ src/game.c
@@ -1,0 +1,279 @@
+#define _XOPEN_SOURCE_EXTENDED // required for UTF-8 encoding
+#include <ncursesw/curses.h> // required for UTF-8 encoding
+#include <locale.h> // required for UTF-8 encoding
+#include <stdlib.h> // required for malloc
+#include <time.h> // required for rand and srand functions to create random numbers
+
+typedef struct Position { // defines a struct for a coordinate position
+  int x;
+  int y;
+} Position;
+
+typedef struct Room { // defines a struct for Rooms 
+  Position position;
+  int height;
+  int width;
+  Position ** doors; // array of pointers to the Position struct type to keep track of where the doors are in a given room
+  // Monster ** monsters;
+  // Item ** items;
+} Room;
+
+typedef struct Level { // defines a struct for a level
+  Room ** rooms; // an array of pointers that will hold all of the instances of the Room struct associated with a particular level
+  char levelMap; // going to hold an array of characters, that will be generated by mapSetup function that will be used to draw the level and refresh when characters, npc's and other entities move
+} Level;
+
+typedef struct Player { // defines a struct for Players 
+  Position position;
+  int health;
+  int attack;
+  int defense;
+  // Room * room;
+} Player;
+
+typedef struct Monster { // defines a struct for monster npcs
+  Position position;
+  int health;
+  int attack;
+  int defense;
+} Monster;
+
+int screenSetup();
+Room ** mapSetup();
+Player * playerSetup();
+int handleInput(int input, Player * user);
+int playerMove(int y, int x, Player * user);
+int checkPosition();
+
+// room creation function definitions
+// Room * createRoom(int x, int y, int height, int width);
+Room * createRoom(int x, int y, int height, int width);
+int drawRoom();
+int connectDoors(Position * doorOne, Position * doorTwo);
+
+int main () {
+	setlocale(LC_ALL, ""); // required for UTF-8 encoding, MUST BE CALLED BEFORE ANYTHING ELSE!
+  srand(time(NULL)); // seeds random number generator with time from the operating system
+  screenSetup(); // initializes ncurses terminal screen
+  mapSetup(); // creates rooms
+  Player * user; // variable that will hold the new user
+	int ch; // variable used to get input from terminal for use in gameloop and handleInput functions
+  user = playerSetup();
+	while ((ch = getch()) != 'q') {
+		handleInput(ch, user);
+	}
+  getch();
+  endwin(); // required for ncurses to work
+  return 0;
+}
+
+int screenSetup() { // initializes ncurses terminal screen 
+  initscr(); // required for ncurses to work
+  printw("Setting up screen.");
+  noecho();
+  refresh();
+  return 0;
+}
+
+Room ** mapSetup () { // this function generates the map for a level and draws it to the ncurses screen and also returns an array of rooms that will be used later for a level struct
+	Room ** rooms; // initializes an array of rooms of the Room struct type
+  rooms = malloc(sizeof(Room)*6); // creates enough memory to store 6 rooms in the array rooms
+  // mvprintw(13, 13, "--------------------");
+  // mvprintw(14, 13, "|..................|");
+  // mvprintw(15, 13, "|..................|");
+  // mvprintw(16, 13, "|..................|");
+  // mvprintw(17, 13, "|..................|");
+  // mvprintw(18, 13, "--------------------");
+
+  rooms[0] = createRoom(13, 13, 6, 8); // creates a room of the given size with the createRoom function and places it in the 0th index of the rooms array
+  drawRoom(rooms[0]); // this function draws the room struct that was created and placed at the 0th index of the rooms array on the ncurses screen
+  
+  // mvprintw(2, 44, "--------------------");
+  // mvprintw(3, 44, "|..................|");
+  // mvprintw(4, 44, "|..................|");
+  // mvprintw(5, 44, "|..................|");
+  // mvprintw(6, 44, "|..................|");
+  // mvprintw(7, 44, "--------------------");
+
+  rooms[1] = createRoom(40, 2, 6, 8); // see line 81
+  drawRoom(rooms[1]); // see line 82
+
+  // mvprintw(10, 40, "--------------------");
+  // mvprintw(11, 40, "|..................|");
+  // mvprintw(12, 40, "|..................|");
+  // mvprintw(13, 40, "|..................|");
+  // mvprintw(14, 40, "|..................|");
+  // mvprintw(15, 40, "--------------------");
+
+  rooms[2] = createRoom(40, 10, 6, 12); // see line 81
+  drawRoom(rooms[2]); // see line 82
+
+  connectDoors(rooms[0]->doors[3], rooms[2]->doors[1]);
+  connectDoors(rooms[1]->doors[2], rooms[0]->doors[0]);
+  // connectDoors(rooms[0]->doors[3], rooms[2]->doors[1]);
+
+  return rooms; // returns the rooms array of type Room struct
+}
+
+Player * playerSetup() {
+  Player * newPlayer; // creates a variable that will hold the yet uncreated new player of the struct type Player
+  newPlayer = malloc(sizeof(Player)); // allocates enough memory to hold a player of type struct Player
+  newPlayer->position.x = 14; // sets starting x position in newPlayer x position member
+  newPlayer->position.y = 14; // sets starting x position in newPlayer y position member
+  newPlayer->health = 20; // sets player starting health
+  playerMove(16, 16, newPlayer); // player start position
+  return newPlayer;
+}
+
+int handleInput(int input, Player * user) { // this function takes the input character from the ch variable in the main method, creates a newX newY set of coordinates based on that input character and then call checkPosition function and pass the new coordinates into that. If the new coordinates are valid, then the checkPosition function will call the playerMove function using the new coordinates
+  int newY;
+  int newX;
+  switch (input) {
+    case 'w':
+      newY = user->position.y - 1;
+      newX = user->position.x;
+      break;
+    case 'a':
+      newY = user->position.y;
+      newX = user->position.x - 1;
+      break;
+    case 's':
+      newY = user->position.y + 1;
+      newX = user->position.x;
+      break;
+    case 'd':
+      newY = user->position.y;
+      newX = user->position.x + 1;
+      break;
+    default:
+      break;
+  }
+  checkPosition(newY, newX, user); // checkPosition function is called with the new coordinates created by the handleInput function
+}
+
+int checkPosition(int newY, int newX, Player * user) { // Check what character is at next intended move position(y and x coordinates), if that word is a period, move the player using playerMove function, else do not move player but move cursor back to position over player @ symbol
+  int space;
+  switch (mvinch(newY, newX)) { // this switch statement is evaluating the character at the next intended Player position using the mvinch ncurses function with the next intened y and x coordinates passed in. If  mvinch returns a . character then the playerMove function is run with the newY, newX positions and the user pointer, pointing to the Player struct passed in
+    case '.':
+      playerMove(newY, newX, user);
+      break;
+		case '+':
+      playerMove(newY, newX, user);
+      break;
+    case '#':
+      playerMove(newY, newX, user);
+      break;
+    default:
+      move(user->position.y, user->position.x); // this case is used to move the cursor back onto the Player if when a user types in a command, the mvinch function determines that the intened move coordinates contain a character that the Player cannot move onto
+      break;
+  }
+}
+
+int playerMove(int newY, int newX, Player * user) { // this command is run to move, the Player(or user) to their new intended move coordinates, whose direction and distance are determined by the switch in handleInput as well as the newY and newX variables being assigned in such a was as to increase the characters position 1 unit in the selected direction as controlled by the WASD keys player movement is mainly moved by this and is constrained by the checkposition function that will stop a player from being moved to a given postion if the tile does not allow a player to move there as set in the checkPosition function
+  mvprintw(user->position.y, user->position.x, "."); // first replace the player's current position with a . symbol
+  user->position.y = newY; // set passed in user's new y position
+  user->position.x = newX; // set passed in user's new x position
+
+  mvprintw(user->position.y, user->position.x, "@"); // print @ symbol at the user's new position defined inside the user struct
+  move(user->position.y, user->position.x); // finally move cursor back on top of the player's @ character by using the position data in the passed in user's struct
+}
+
+Room * createRoom(int x, int y, int height, int width) { // this creates a room with the specified position and size, then generates four doors in random location along the room's four sides
+  Room * newRoom;
+  newRoom = malloc(sizeof(Room));
+  newRoom->position.x = x;
+  newRoom->position.y = y;
+  newRoom->height = height;
+  newRoom->width = width;
+
+  newRoom->doors = malloc(sizeof(Position) * 4);
+
+  // top door
+  newRoom->doors[0] = malloc(sizeof(Position));
+  newRoom->doors[0]->x = rand() % (width - 2) + newRoom->position.x + 1;
+  newRoom->doors[0]->y = newRoom->position.y;
+
+  // left door
+  newRoom->doors[1] = malloc(sizeof(Position));
+  newRoom->doors[1]->y = rand() % (height - 2) + newRoom->position.y + 1;
+  newRoom->doors[1]->x = newRoom->position.x;
+
+  // bottom door
+  newRoom->doors[2] = malloc(sizeof(Position));
+  newRoom->doors[2]->x = rand() % (width - 2) + newRoom->position.x + 1;
+  newRoom->doors[2]->y = newRoom->position.y + newRoom->height - 1;
+
+  // right door
+  newRoom->doors[3] = malloc(sizeof(Position));
+  newRoom->doors[3]->y= rand() % (height - 2) + newRoom->position.y + 1;
+	newRoom->doors[3]->x = newRoom->position.x + width - 1;
+  return newRoom;
+}
+
+int drawRoom(Room * room) { // draws a room on the ncurses screen
+  int x;
+  int y;
+  // draw top and bottom of a room on the ncurses screen
+  for (x = room->position.x; x < room->position.x + room->width; x++) {
+    mvprintw(room->position.y, x, "-"); // drawing top of room
+    mvprintw(room->position.y + room->height - 1, x, "-"); // drawing bottom of room
+  }
+
+  // drawing walls here on the ncurses screen
+  for (y = room->position.y + 1; y < room->position.y + room->height - 1; y++) {
+    mvprintw(y, room->position.x, "|");
+    mvprintw(y, room->position.x + room->width - 1, "|");
+    // drawing floors in between walls here on the ncurses screen
+    for (x = room->position.x + 1; x < room->position.x + room->width - 1; x++) {
+      mvprintw(y, x, ".");
+    }
+  }
+
+  // draw doors here on the ncurses screen
+  mvprintw(room->doors[0]->y, room->doors[0]->x, "+" );
+  mvprintw(room->doors[1]->y, room->doors[1]->x, "+" );
+  mvprintw(room->doors[2]->y, room->doors[2]->x, "+" );
+  mvprintw(room->doors[3]->y, room->doors[3]->x, "+" );
+  return 1;
+}
+
+int connectDoors(Position * doorOne, Position * doorTwo) {
+  int count = 0;
+  Position temp;
+  Position previous;
+  temp.x = doorOne->x;
+  temp.y = doorOne->y;
+  previous = temp;
+
+  while(1) {
+    // step left
+    if ((abs((temp.x - 1) - doorTwo->x ) < abs(temp.x - doorTwo->x)) && (mvinch(temp.y, temp.x - 1) == ' ')) {
+      previous.x = temp.x;
+      temp.x = temp.x - 1;
+      // step right
+    } else if ((abs((temp.x + 1) - doorTwo->x ) < abs(temp.x - doorTwo->x)) && (mvinch(temp.y, temp.x + 1) == ' ')) {
+      previous.x = temp.x;
+      temp.x = temp.x + 1;
+      // step down
+    } else if ((abs((temp.y + 1) - doorTwo->y ) < abs(temp.y - doorTwo->y)) && (mvinch(temp.y + 1, temp.x) == ' ')) {
+      previous.y = temp.y;
+      temp.y = temp.y + 1;
+      // step up
+    } else if ((abs((temp.y - 1) - doorTwo->y ) < abs(temp.y - doorTwo->y)) && (mvinch(temp.y - 1, temp.x) == ' ')) {
+      previous.y = temp.y;
+      temp.y = temp.y - 1;
+    } else {
+      if (count == 0) {
+        temp = previous;
+        count++;
+        continue;
+      } else {
+        return 0;
+      }
+    }
+    mvprintw(temp.y, temp.x, "#");
+    // getch();
+  }
+
+  return 1;
+}
